requirements.txt
argparse
qrcode
pillow 